requirements.txt

Django==4.1.5
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-extensions==3.2.1
django-filter==22.1
django-js-asset==2.0.0
django-storages==1.13.2
djangorestframework==3.14.0

psycopg2==2.9.5